requirements.txt
pandas
numpy
nltk.
nltk.stem
nltk.corpus
nltk.tokenize
sklearn
re
pickle
scipy.io
